Revert "run graalvm native tests github action"
This reverts commit 5a78366b
